connection to db send to config file
